// src/constants/types.ts
export enum PokemonTypes {
    normal = 'normal',
    fire = 'fire',
    water = 'water',
    electric = 'electric',
    grass = 'grass',
    ice = 'ice',
    fighting = 'fighting',
    poison = 'poison',
    ground = 'ground',
    flying = 'flying',
    psychic = 'psychic',
    ghost = 'ghost',
    shadow = 'shadow',
    fairy = 'fairy',
    dragon = 'dragon',
    dark = 'dark',
    steel ='steel',
    unknown = 'unknown',
    bug = 'bug',
    rock = 'rock'
}